fix: перенес функцию вычисления, в types.js
