+++ src/pages/PostPageVuex.vue
@@ -51,7 +51,7 @@ import MyDialog from '@/components/UI/MyDialog.vue';
 import MySelect from '@/components/UI/MySelect.vue';
 import MyInput from '@/components/UI/MyInput.vue';
 import store from '@/store/index.js';
-import STATE_TYPES from '@/store/Types.js';
+import { STATE_TYPES } from '@/store/Types.js';
 import { mapState, mapGetters, mapActions, mapMutations } from 'vuex';
 export default {
   components: {
